+++ CHANGES
@@ -4,6 +4,10 @@
 
 = GENI Portal Release Notes =
 
+== 3.1 ==
+ * Deploy GENI/ORBIT sync service (geni-sync-wireless.py) and change
+   Wireless Account Management page (wimax-enable.php) accordinlgy (#1392).
+
 == 3.0 ==
  * Add create image function for PG/IG racks within jacks-app (#1389).
  * Remove account and identity (and associated) tables, scripts and clients (#299)
